chore: load fonts with next/font
