+++ Frontend/src/app/auth/social/page.tsx
@@ -3,6 +3,7 @@
 import { Suspense } from "react";
 import SocialAuthHandler from "./SocialAuthHandler";
 
+
 export default function SocialAuthPage() {
   return (
     <Suspense>
