+++ Frontend/src/app/layout.tsx
@@ -1,9 +1,16 @@
 import type { Metadata } from "next";
+import { Inter, Roboto_Mono } from "next/font/google";
 import { AuthProvider } from "../contexts/AuthContext";
 import { VehicleProvider } from "../contexts/VehicleContext";
 import "./globals.css";
 import NavBar from "@/components/NavBar";
 import OfflineTripSync from "@/components/OfflineTripSync";
+
+const inter = Inter({ subsets: ["latin"], variable: "--font-geist-sans" });
+const robotoMono = Roboto_Mono({
+        subsets: ["latin"],
+        variable: "--font-geist-mono",
+});
 
 
 export const metadata: Metadata = {
@@ -13,13 +20,13 @@ export const metadata: Metadata = {
 };
 
 export default function RootLayout({
-	children,
+        children,
 }: Readonly<{
-	children: React.ReactNode;
+        children: React.ReactNode;
 }>) {
-	return (
-		<html lang="es">
-			<body>
+        return (
+                <html lang="es" className={`${inter.variable} ${robotoMono.variable}`}>
+                        <body>
                                 <AuthProvider>
                                         <VehicleProvider>
                                                 <NavBar />
